+++ src/PluginManager/include/PluginManager/SkyConnectIntf.h
@@ -48,8 +48,9 @@ public:
     enum struct RecordingMode {
         /*! A (new) flight with a single aircrat is to be recorded. */
         SingleAircraft,
-        /*! The aircraft is to be added to the current flight; existing aircraft
-         *  are replayed during recording.
+        /*!
+         * The aircraft is to be added to the current flight; existing aircraft
+         * are replayed during recording.
          */
         AddToFormation
     };
@@ -72,7 +73,19 @@ public:
     virtual ReplayMode getReplayMode() const noexcept = 0;
     virtual void setReplayMode(ReplayMode replayMode) noexcept = 0;
 
-    virtual void startRecording(RecordingMode recordingMode, const InitialPosition &initialPosition = InitialPosition()) noexcept = 0;
+    /*!
+     * Starts recording the flight. Depending on the \C recordingMode already recorded formation aircraft
+     * are replayed during recording. If the \c initialPosition is given (\c isNull() returns false) then
+     * the user aircraft is placed at the given \c initialPosition before recording. This position is
+     * typically calculated to be relative of the previous user aircraft in the formation.
+     *
+     * \param recordingMode
+     *        the RecordingMode which controls whether existing formation aircraft are to be replayed
+     * \param initialPosition
+     *        the optional initial position where the current user aircraft is placed before recording;
+     *        set to a \e null position if the user aircraft should keep its current initial position
+     */
+    virtual void startRecording(RecordingMode recordingMode, const InitialPosition &initialPosition = InitialPosition::NullData) noexcept = 0;
     virtual void stopRecording() noexcept = 0;
 
     /*!
@@ -98,8 +111,7 @@ public:
      */
     virtual bool isInRecordingState() const noexcept = 0;
 
-
-    virtual void startReplay(bool fromStart, const InitialPosition &flyWithFormationPosition = InitialPosition()) noexcept = 0;
+    virtual void startReplay(bool fromStart, const InitialPosition &flyWithFormationPosition = InitialPosition::NullData) noexcept = 0;
     virtual void stopReplay() noexcept = 0;
 
     /*!